LayoutTests/sputnik/Conformance/12_Statement/12.2_Variable_Statement/S12.2_A8_T5-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token '--'. Expected a parameter pattern or a ')' in parameter list.
S12.2_A8_T5

PASS Expected parsing failure

TEST COMPLETE
